Refactor compiler into modules and expand CI checks
