+++ lockstep_compiler/__init__.py
@@ -1,0 +1,19 @@
+from .cli import build_arg_parser, run_cli
+from .compiler import compile_lockstep
+from .errors import LockstepCompileError, ParseErrorCollector
+from .models import LockstepCompileResult, LockstepDiagnostic, normalize_diagnostics
+from .visitors import build_debug_visitor, build_semantic_validator, validate_semantics
+
+__all__ = [
+    "LockstepCompileError",
+    "LockstepCompileResult",
+    "LockstepDiagnostic",
+    "ParseErrorCollector",
+    "build_arg_parser",
+    "build_debug_visitor",
+    "build_semantic_validator",
+    "compile_lockstep",
+    "normalize_diagnostics",
+    "run_cli",
+    "validate_semantics",
+]
