+++ lockstep_compiler/cli.py
@@ -1,0 +1,60 @@
+import argparse
+import sys
+from pathlib import Path
+
+from .errors import LockstepCompileError
+
+
+def build_arg_parser():
+    parser = argparse.ArgumentParser(
+        description="Debug parser for Lockstep source files."
+    )
+    parser.add_argument(
+        "path",
+        nargs="?",
+        help="Optional path to a Lockstep source file. Reads from stdin when omitted.",
+    )
+    return parser
+
+
+def run_cli(argv=None, *, stdin=None, stderr=None, compiler=None):
+    parser = build_arg_parser()
+    args = parser.parse_args(argv)
+
+    stdin = sys.stdin if stdin is None else stdin
+    stderr = sys.stderr if stderr is None else stderr
+
+    if args.path:
+        source_path = Path(args.path)
+        try:
+            source = source_path.read_text(encoding="utf-8")
+        except FileNotFoundError:
+            print(f"Unable to read '{source_path}': file not found.", file=stderr)
+            return 1
+        except PermissionError as err:
+            reason = err.strerror or "permission denied"
+            print(f"Unable to read '{source_path}': {reason}.", file=stderr)
+            return 1
+        except UnicodeDecodeError as err:
+            print(
+                f"Unable to read '{source_path}': invalid UTF-8 ({err.reason}).",
+                file=stderr,
+            )
+            return 1
+    else:
+        source = stdin.read()
+
+    try:
+        compiler(source)
+    except LockstepCompileError as err:
+        count = len(err.errors)
+        suffix = "" if count == 1 else "s"
+        print(
+            f"Compilation failed with {count} {err.phase} error{suffix}.",
+            file=stderr,
+        )
+        for error in err.errors:
+            print(f"line {error.line}:{error.column} {error.message}", file=stderr)
+        return 1
+
+    return 0
