+++ pyproject.toml
@@ -28,3 +28,11 @@ omit = [
 
 [tool.coverage.report]
 show_missing = true
+
+[build-system]
+requires = ["setuptools>=68", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+py-modules = ["debug_compiler"]
+packages = ["lockstep_compiler"]
